pyproject.toml
[tool.poetry]
name = "leaf"
version = "0.1.0"
description = "LabEquipment Adapter Framework (LEAF)"
authors = [
    "Matthew Crowther <b5044420@newcastle.ac.uk>",
    "Brett Metcalfe <brett.metcalfe@hotmail.co.uk>",
    "Bart Nijsse <bart.nijsse@wur.nl>",
    "Jasper Koehorst <jasper.koehorst@wur.nl>"
]
license = "Apache-2.0"
readme = "README.md"
repository = "https://gitlab.com/LabEquipmentAdapterFramework/leaf"
homepage = "https://gitlab.com/LabEquipmentAdapterFramework/leaf"
documentation = "https://gitlab.com/LabEquipmentAdapterFramework/leaf/issues"

[tool.poetry.dependencies]
python = "^3.6"
setuptools = "~75.1.0"
paho-mqtt = "~2.1.0"
redis = "~5.0.0b2"
aioredis = "~1.3.1"
influxobject = "~0.0.1"
pyyaml = "~6.0.2"
poetry = "~1.8.3"
watchdog = "~2.1.6"
# mypy = "~0.910.0"
pandas = "~2.2.3"
dateparser = "~1.2.0"

[tool.poetry.scripts]
leaf = "leaf.start:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
coverage = "^7.0"
